fix: userId, eventId의 타입을 Long에서 int로 수정

+++ src/main/java/ureca/ureca_mini/winner/EntryRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class EntryRequest {
 
-    private Long userId;
-    private Long eventId;
+    private int userId;
+    private int eventId;
 }
